~ não sobrescrever config onedrive + function log ~ corrigido nome arquivos de log ! testar e implementar funcion de log futuramente

+++ scripts/4-onedrive-config.sh
@@ -1,11 +1,32 @@
+#!/bin/bash
+
 #https://github.com/skilion/onedrive
+
+# Carimbar arquivo de log
+timestamp=$(date +%Y-%m-%d_%H-%M-%S)
+echo "onedrive-config: $timestamp" >> $HOME/Downloads/SicoobInstala.log
+
+# Verificar caminhos das dependencias
+if [ "$scriptsDir" == "" ] && [ -e "./scripts/" ];then
+    scriptsDir=./scripts
+    relativeDir=.
+elif [ "$scriptsDir" == "" ];then
+    scriptsDir=.
+    relativeDir=..
+fi
+
+# Dependendencias
+source $scriptsDir/_GeneralFunctions.sh
 
 onedrive
 
 # Copiar arquivo de configuração padrão
-cp /usr/share/doc/onedrive/config ~/.config/onedrive/config
+if ! [ -e "$HOME/.config/onedrive/config" ]; then 
+    cp /usr/share/doc/onedrive/config ~/.config/onedrive/config
+fi
 
 # Mostrar configuração atual
+echo -e "Abaixo configuração atual do Onedrive\n"
 onedrive --display-config
 
 # Configure
